requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
PyYAML
google-genai
google-auth
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-grpc
